% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_id.R
\name{anonymize_id}
\alias{anonymize_id}
\title{Anonymize ID}
\usage{
anonymize_id(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Create anonymized ID from any vector
}
\examples{
anonymize_id(c("b", "a", "d", "a"))

}
