% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lines.R
\name{count_lines}
\alias{count_lines}
\title{Count lines in R files}
\usage{
count_lines(path = ".")
}
\arguments{
\item{path}{directory path}
}
\description{
Count lines in all R files in a certain path.
}
\examples{
count_lines()

}
