% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_c.R
\name{drop_c}
\alias{drop_c}
\title{Drop column from data.frame}
\usage{
drop_c(df, columns)
}
\arguments{
\item{df}{A data.frame}

\item{columns}{A character of vector of characters representing column names}
}
\description{
Drops specified column from data.frame
}
\author{
Johan Junkka
}

