% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_t.R
\name{freq_t}
\alias{freq_t}
\title{Frequency table}
\usage{
freq_t(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Generate a frequency table from a vector, containing 
frequencies, cumulative frequency and proportional.
}
\examples{
x <- sample(LETTERS[1:10], 44, TRUE)
freq_t(x)

}
\author{
Johan Junkka \email{johan.junkka@gmail.com}
}
