% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ghelp.R
\name{ghelp}
\alias{ghelp}
\title{GitHub help}
\usage{
ghelp(topic, in_cran = TRUE)
}
\arguments{
\item{topic}{character. A search string.}

\item{in_cran}{logical. If TRUE the search is limited to CRAN
repositories}
}
\description{
Search GitHub for R code examples.
}
\examples{
ghelp('vapply')
}

