% Generated by roxygen2 (4.0.0): do not edit by hand
\name{knitHTML}
\alias{knitHTML}
\title{Knit HTML with template}
\usage{
knitHTML(file, title = file, homeD = homeD, preview = TRUE)
}
\arguments{
  \item{file}{Name of .Rmd file}

  \item{title}{Title in meta}

  \item{homeD}{Name of root dirctory of project}

  \item{preview}{Logical, if true will show html in Google
  Chrome}
}
\description{
Compile .Rmd file to html with custom bootstrap template. Naming convention filename/filename.Rmd
}

