% Generated by roxygen2 (4.0.0): do not edit by hand
\name{knitPDF}
\alias{knitPDF}
\title{Knit PDF with Pandoc}
\usage{
knitPDF(file, homeD, preview = TRUE)
}
\arguments{
  \item{file}{Name of .Rmd file}

  \item{homeD}{Name of root dirctory of project}

  \item{preview}{Logical, if true will show html in Google
  Chrome}
}
\description{
Compile .Rmd file to .pdf with knitr and pandoc. Naming convention filename/filename.Rmd
}

