% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ll.R
\name{ll}
\alias{ll}
\title{List objects in environment}
\source{
\url{http://stackoverflow.com/questions/1358003}
}
\usage{
ll(..., n = 10)
}
\arguments{
\item{...}{pass to \code{.ls.object}}

\item{n}{number of objects to display}
}
\description{
With size and dimensions
}

