% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_unique.R
\name{n_unique}
\alias{n_unique}
\title{Count unique}
\usage{
n_unique(x)
}
\arguments{
\item{x}{a vector or data.frame}
}
\description{
Count unique entries in a vector of in all columns of a
data.frame
}
\examples{
x <- data.frame(
 'a' = rep(1:4, 2),
 'b' = rep(1:2, 4)
)
n_unique(x)
n_unique(x$a)

}
\author{
Johan Junkka \email{johan.junkka@gmail.com}
}

