% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas.R
\name{nas}
\alias{nas}
\title{Show NA}
\usage{
nas(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A table with counts of NA TRUE and FALSE
}
\description{
Get counts and proportions of NA and non NA in vector
}
\author{
Johan Junkka
}
