% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nas_c}
\alias{nas_c}
\title{Proportion NA in data.frame by column}
\usage{
nas_c(x)
}
\arguments{
\item{x}{a data.frame}
}
\description{
Calculates proportion of NA values by column in a data.frame
}
\examples{
x <- data.frame(
  'a' = c(1:180, rep(NA, 20)),
  'b' = c(1:160, rep(NA, 40)),
  'c' = c(1:200)
)
nas_c(x)
}
\author{
Johan Junkka \email{johan.junkka@gmail.com}
}

