% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sp_to_json}
\alias{sp_to_json}
\title{Spatial object to geojson}
\usage{
sp_to_json(map, name = "map")
}
\arguments{
\item{map}{Spatial object}

\item{name}{character string of the name of the object}
}
\description{
Creates a geojson file of a \code{\link[sp]{SpatialPolygons}} or
  \code{\link[sp]{SpatialPoints}} object.
}
\examples{
data(nor2k, package = "rgdal")
sp_to_json(nor2k, "nor2k")
}
\author{
Johan Junkka <johan.junkka@gmail.com>
}
\seealso{
\link[sp]{SpatialPolygons} \link[sp]{SpatialPoints}
}

