% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drop_c}
\alias{drop_c}
\title{Drop column from data.frame}
\usage{
drop_c(df, columns)
}
\arguments{
\item{df}{A data.frame}

\item{columns}{A character of vector of characters representing column names}
}
\description{
Drops specified column from data.frame
}
\author{
Johan Junkka
}

