% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ll}
\alias{ll}
\title{List Objects in Environment}
\usage{
ll(order_by = "class")
}
\arguments{
\item{order_by}{= "class" Order by "class" or "object"}
}
\value{
A matrix containing name and class of objects
}
\description{
Shows objects name and class in local environment. An extention of base::ls.
}
\author{
Johan Junkka
}

